+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.7.0
 + Use greedy integrators where possible (Fixes #57)
++ Add [JSpecify](https://jspecify.dev/) annotations for static analysis
 
 ### 0.6.0
 + Implement `dropLast(n)`